Remove TnT Countdown Bukkit implementation
