+++ bukkit-example-json/src/main/java/com/lunarclient/apollo/example/json/module/TntCountdownJsonExample.java
@@ -28,13 +28,9 @@ import com.lunarclient.apollo.example.ApolloExamplePlugin;
 import com.lunarclient.apollo.example.json.util.JsonPacketUtil;
 import com.lunarclient.apollo.example.json.util.JsonUtil;
 import com.lunarclient.apollo.example.module.impl.TntCountdownExample;
-import java.lang.reflect.Method;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.UUID;
 import org.bukkit.Bukkit;
-import org.bukkit.Location;
-import org.bukkit.World;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.Player;
 import org.bukkit.entity.TNTPrimed;
@@ -44,16 +40,6 @@ import org.bukkit.event.Listener;
 import org.bukkit.event.entity.EntitySpawnEvent;
 
 public class TntCountdownJsonExample extends TntCountdownExample implements Listener {
-
-    private static Method entityGetter;
-
-    static {
-        try {
-            TntCountdownJsonExample.entityGetter = Bukkit.class.getDeclaredMethod("getEntity", UUID.class);
-        } catch (Throwable throwable) {
-            // Ignore for legacy versions.
-        }
-    }
 
     public TntCountdownJsonExample() {
         Bukkit.getPluginManager().registerEvents(this, ApolloExamplePlugin.getInstance());
@@ -70,31 +56,10 @@ public class TntCountdownJsonExample extends TntCountdownExample implements List
 
     @Override
     public void overrideTntCountdownExample(Player viewer) {
-        Location location = viewer.getLocation();
-        TNTPrimed entity = viewer.getWorld().spawn(location, TNTPrimed.class);
         int customTicks = 200;
 
-        TNTPrimed target = null;
-        if (TntCountdownJsonExample.entityGetter != null) {
-            try {
-                target = (TNTPrimed) TntCountdownJsonExample.entityGetter.invoke(null, entity.getUniqueId());
-            } catch (Throwable throwable) {
-                throwable.printStackTrace();
-            }
-        } else {
-            for (World world : Bukkit.getWorlds()) {
-                for (TNTPrimed compare : world.getEntitiesByClass(TNTPrimed.class)) {
-                    if (compare.getUniqueId().equals(entity.getUniqueId())) {
-                        target = compare;
-                        break;
-                    }
-                }
-            }
-        }
-
-        if (target != null) {
-            target.setFuseTicks(customTicks);
-        }
+        TNTPrimed entity = viewer.getWorld().spawn(viewer.getLocation(), TNTPrimed.class);
+        entity.setFuseTicks(customTicks);
 
         JsonPacketUtil.sendPacket(viewer, this.createTNTCountdownMessage(entity, customTicks));
     }
